Fix failing cusparse 12.4 tests (#484)
* Disable bad args tests for spmat, dnmar, spvec, dnvec routines. Also dont use mixed index precision in sparse to dense and dense to sparse as not supported by cusparse

* Clang formatting

* Undo code change

* Do not test invalid arguments with cusparse backend for pmat, dnmat, spvec, and dnvec

* Move #endif after INSTANTIATE_TEST_SUITE_P in test_xxx.cpp files

* Clang formatting

* Move #endif after INSTANTIATE_TEST_SUITE_P in test_gthr.cpp file

* Move #endif after INSTANTIATE_TEST_SUITE_P in test_csr2dense.cpp, test_dense2csr.cpp, and test_dense2csc.cpp files

* Clang formatting

* Fix dense_to_sparse_csr_bad_args failing test

---------


[ROCm/hipSPARSE commit: 4d19a656835bb04decccfb0f26a5e937488a5210]

+++ clients/include/testing_dense_to_sparse_coo.hpp
@@ -59,7 +59,7 @@ void testing_dense_to_sparse_coo_bad_arg(void)
     auto ddense_val_managed
         = hipsparse_unique_ptr{device_malloc(sizeof(float) * safe_size), device_free};
     auto dcoo_row_ind_managed
-        = hipsparse_unique_ptr{device_malloc(sizeof(int64_t) * safe_size), device_free};
+        = hipsparse_unique_ptr{device_malloc(sizeof(int32_t) * safe_size), device_free};
     auto dcoo_col_ind_managed
         = hipsparse_unique_ptr{device_malloc(sizeof(int32_t) * safe_size), device_free};
     auto dcoo_val_managed
@@ -67,7 +67,7 @@ void testing_dense_to_sparse_coo_bad_arg(void)
     auto dbuf_managed = hipsparse_unique_ptr{device_malloc(sizeof(char) * safe_size), device_free};
 
     float*   ddense_val   = (float*)ddense_val_managed.get();
-    int64_t* dcoo_row_ind = (int64_t*)dcoo_row_ind_managed.get();
+    int32_t* dcoo_row_ind = (int32_t*)dcoo_row_ind_managed.get();
     int32_t* dcoo_col_ind = (int32_t*)dcoo_col_ind_managed.get();
     float*   dcoo_val     = (float*)dcoo_val_managed.get();
     void*    dbuf         = (void*)dbuf_managed.get();
